requirements.txt
cycler==0.10.0
matplotlib==2.1.0
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
scipy==0.19.1
six==1.11.0
